requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==5.1.1
django-ckeditor==6.7.1
django-cors-headers==4.4.0
django-js-asset==2.2.0
django-modeltranslation==0.19.9
django-unfold==0.39.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
pillow==10.4.0
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
